Update
Reclassified main running thing, but properly this time.

Main generation class is for 1 turn + there's a separate class for running multiple turns

Most of playground.py moved to generation.py to leave that script to handle more high level operations and the like.

Work will need to be done to turn this into something useful + modular. That's probably the next big step forwards after this.