Testing
I want to return to this project, started with a new domain yet again, it's very broken, idea's to get a core turn thing setup that manages all the stuff that has to be done every turn. I have no idea where I'm gonna go with this project, I just wanted to work on it some.